+++ .gitignore
@@ -1,1 +1,4 @@
 /target/
+.settings/
+result/
+test-output/
