--- src/source/Core/PageObject.java
@@ -1,5 +1,0 @@
-package Core;
-
-public class PageObject {
-
-}
