+++ src/test/java/tests/NewTest_1.java
@@ -2,9 +2,7 @@ package tests;
 
 import org.testng.annotations.Test;
 
-import Pages.BasePage;
-
-public class NewTest_1 extends BasePage{
+public class NewTest_1{
   @Test
   public void f() {
 	  System.out.print("HELLO");
